ref: remove mock file in testing
