+++ contracts/os/andromeda-ibc-registry/src/testing/mod.rs
@@ -1,2 +1,1 @@
-mod mock;
 mod tests;
